+++ e2e/playwright.config.ts
@@ -4,7 +4,7 @@ import { defineConfig, devices } from '@playwright/test'
 
 const isCI = process.env['CI']
 
-const baseURL = process.env['BASE_URL'] ?? 'http://localhost:6006'
+const baseURL = 'http://localhost:5173'
 
 const times = {
   '1min': 60 * 1000,
@@ -12,12 +12,19 @@ const times = {
 }
 
 export default defineConfig({
+  webServer: {
+    // it will start web app before running tests so you don't need to start it manually
+    command: 'pnpm run start',
+    url: baseURL,
+    stdout: 'ignore',
+    stderr: 'pipe',
+  },
   testDir: './tests',
   fullyParallel: true,
   forbidOnly: !!isCI,
   retries: isCI ? 2 : 0,
   workers: isCI ? 1 : undefined,
-  reporter: 'html',
+  reporter: 'line',
   globalTimeout: isCI ? 5 * times['1min'] : undefined,
   timeout: isCI ? 1 * times['1min'] : undefined,
   use: {