+++ packages/ui/src/components/DateInput/__stories__/Template.tsx
@@ -8,5 +8,5 @@ export const Template: StoryFn<typeof DateInput> = props => (
 Template.args = {
   label: 'Date Input',
   required: true,
-  placeholder: 'YYYY-MM-DD',
+  placeholder: 'MM-DD-YYYY',
 }