+++ packages/ui/src/components/DateInput/components/CalendarDaily.tsx
@@ -14,17 +14,30 @@ import {
   isSameDay,
 } from '../helpers'
 
-const ButtonDate = styled(Button)`
+const Day = styled(Button)`
   height: ${({ theme }) => theme.sizing['312']};
   width: 100%;
   padding: 0;
-`
+  color: ${({ theme }) => theme.colors.neutral.textWeak};
 
-const RangeButton = styled(Button)`
-  background-color: ${({ theme }) => theme.colors.primary.background};
-  height: ${({ theme }) => theme.sizing['312']};
-  width: 100%;
-  padding: 0;
+  &[aria-label="in-range"] {
+    color:  ${({ theme }) => theme.colors.primary.textHover};
+    background-color: ${({ theme }) => theme.colors.primary.background};
+  }
+
+  &[aria-label="in-range"]:hover {
+    color: ${({ theme }) => theme.colors.neutral.textStronger};
+    background-color: ${({ theme }) => theme.colors.primary.backgroundStrongHover};
+  }
+
+  &[aria-label="not-current"], :disabled {
+    color: ${({ theme }) => theme.colors.neutral.textDisabled};
+  }
+
+  &[aria-label="selected"] {
+    color: ${({ theme }) => theme.colors.neutral.textStronger};
+  }
+
 `
 
 const CapitalizedText = styled(Text)`
@@ -36,7 +49,7 @@ const CapitalizedText = styled(Text)`
   }
 `
 
-export const Daily = ({ disabled }: { disabled: boolean }) => {
+export const Daily = () => {
   const {
     value,
     yearToShow,
@@ -55,13 +68,15 @@ export const Daily = ({ disabled }: { disabled: boolean }) => {
     setInputValue,
     format,
     setVisible,
+    readOnly,
+    disabled,
+    hoveredDate,
+    setHoveredDate,
   } = useContext(DateInputContext)
 
-  const [rangeState, setRangeState] = useState<'start' | 'none' | 'done'>(
-    range?.start ? 'start' : 'none',
-  ) // Used when selectsRange is True. Kow the current state of the range: none when start date not selected, start when start date is selected, done when start & end date selected
-
-  const [hoveredDate, setHoveredDate] = useState<Date | null>(null)
+  const [rangeState, setRangeState] = useState<'start' | 'end'>(
+    range?.start && !range?.end ? 'end' : 'start',
+  ) // Used when selectsRange is True. The current state of the range: "start" when one must select the start-date of the range, "end" when start date is selected and one must select the end-date
 
   const monthDays = new Date(yearToShow, monthToShow, 0).getDate() // Number of days in the month
 
@@ -113,6 +128,7 @@ export const Daily = ({ disabled }: { disabled: boolean }) => {
           variant="bodyStrong"
           sentiment="neutral"
           key={day[0]}
+          disabled={disabled}
         >
           {day[1]}
         </CapitalizedText>
@@ -149,7 +165,14 @@ export const Daily = ({ disabled }: { disabled: boolean }) => {
             range?.start &&
             constructedDate > range.start &&
             hoveredDate &&
-            constructedDate < hoveredDate) ||
+            constructedDate < hoveredDate &&
+            !range.end) ||
+          (selectsRange &&
+            range?.start &&
+            constructedDate < range.start &&
+            hoveredDate &&
+            constructedDate > hoveredDate &&
+            !range.end) ||
           (range?.start &&
             range.end &&
             constructedDate < range.end &&
@@ -177,24 +200,8 @@ export const Daily = ({ disabled }: { disabled: boolean }) => {
 
         const onClickRange = (event: MouseEventReact, newDate: Date) => {
           if (selectsRange) {
-            // Selecting start date
-            if (rangeState === 'none') {
-              setRange?.({ start: newDate, end: null })
-              onChange?.([newDate, null], event)
-              setInputValue(
-                formatValue(
-                  null,
-                  { start: newDate, end: null },
-                  false,
-                  true,
-                  format,
-                ),
-              )
-              setRangeState('start')
-            }
-
-            // Selecting end date
-            else if (isAfterStartDate) {
+            // Selecting the end date
+            if (rangeState === 'end' && isAfterStartDate) {
               setRange?.({ start: range.start, end: newDate })
               onChange?.([range.start, newDate], event)
               setInputValue(
@@ -206,11 +213,34 @@ export const Daily = ({ disabled }: { disabled: boolean }) => {
                   format,
                 ),
               )
+              // If we click on another date, it will reset the range
+              setRangeState('start')
               setVisible(false)
-              setRangeState('done')
-            } else {
-              // End date before start
+            }
+            // The newDate is before, we swap end & start date
+            else if (
+              rangeState === 'end' &&
+              !isAfterStartDate &&
+              range?.start
+            ) {
+              setRange?.({ start: newDate, end: range.start })
+              onChange?.([newDate, range.start], event)
+              setInputValue(
+                formatValue(
+                  null,
+                  { start: newDate, end: range.start },
+                  false,
+                  true,
+                  format,
+                ),
+              )
+              setRangeState('start')
+              setVisible(false)
+            }
+            // Selecting the start date
+            else {
               setRange?.({ start: newDate, end: null })
+              onChange?.([newDate, null], event)
               setInputValue(
                 formatValue(
                   null,
@@ -220,10 +250,11 @@ export const Daily = ({ disabled }: { disabled: boolean }) => {
                   format,
                 ),
               )
-              onChange?.([newDate, null], event)
+              setRangeState('end')
             }
           }
         }
+
         const createTestId = () => {
           if (isInHoveredRange) return 'rangeButton'
           if (data.month === -1) return 'dayLastMonth'
@@ -232,7 +263,13 @@ export const Daily = ({ disabled }: { disabled: boolean }) => {
           return undefined
         }
 
-        const Day = isInHoveredRange ? RangeButton : ButtonDate
+        const dayState = () => {
+          if (isSelected) return 'selected'
+          if (isInHoveredRange) return 'in-range'
+          if (data.month !== 0) return 'not-current'
+
+          return 'neutral'
+        }
 
         return (
           <Day
@@ -241,7 +278,7 @@ export const Daily = ({ disabled }: { disabled: boolean }) => {
             disabled={disabled || isExcluded || isOutsideRange}
             key={`${data.month}-${data.day}`}
             onClick={event => {
-              if (!isExcluded && !isOutsideRange) {
+              if (!isExcluded && !isOutsideRange && !readOnly) {
                 const newDate = getNewDate()
 
                 if (selectsRange) {
@@ -259,22 +296,10 @@ export const Daily = ({ disabled }: { disabled: boolean }) => {
             onMouseEnter={() => {
               if (selectsRange && range?.start) setHoveredDate(constructedDate)
             }}
-            onMouseLeave={() => {
-              if (selectsRange && range?.start) setHoveredDate(null)
-            }}
+            aria-label={dayState()}
+            data-testid={createTestId()}
           >
-            <Text
-              as="span"
-              variant="bodyStrong"
-              prominence={isSelected && !isInHoveredRange ? 'strong' : 'weak'}
-              sentiment={isSelected || isInHoveredRange ? 'primary' : 'neutral'}
-              disabled={
-                disabled || data.month !== 0 || isExcluded || isOutsideRange
-              }
-              data-testid={createTestId()}
-            >
-              {data.day}
-            </Text>
+            {data.day}
           </Day>
         )
       })}